feat: Add support for SSD1306 OLED display
This commit adds support for the SSD1306 OLED display by including the necessary library and initializing the display in the setup function. It also updates the loop function to display the current time on the OLED screen. This feature enhances the functionality of the project by providing a visual representation of the time.